♻️ refactor: 리뷰 관련 API 수정

+++ src/main/java/com/jajaja/domain/review/service/ReviewCommandServiceImpl.java
@@ -49,7 +49,7 @@ public class ReviewCommandServiceImpl implements ReviewCommandService {
         Review review = Review.builder()
                 .member(member)
                 .product(product)
-                .rating(dto.rating())
+                .rating(dto.rating().byteValue())
                 .content(dto.content())
                 .build();
         Review savedReview = reviewRepository.save(review);
